app/requirements.txt
streamlit==1.38.0
mlflow==2.10.0
pandas==2.2.1
numpy==1.24.4
Pillow==10.3.0
boto3==1.34.44
psutil==5.9.7
scikit-learn==1.4.1.post1
pyyaml==6.0.1
